+++ CHANGELOG.md
@@ -12,6 +12,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Added
 
+- Added Python2 and Python3 versions of generic Python F2PY macros.
+
+## [3.3.9] - 2021-04-14
+
+### Added
+
+- Added option `BUILD_WITH_PFLOGGER` which defaults to `ON`. This is
+  added for collaborators that do not use pFlogger
+
+## [3.3.8] - 2021-04-09
+
+### Added
+
+- Added new `ESMA_USE_GFE_NAMESPACE` which defaults to `OFF`. If you set this to `ON`, you must then use the new GFE namespace style in CMake, e.g., `gftl` ==> `GFTL::gftl`. 
+
 ## [3.3.7] - 2021-03-09
 
 ### Fixed